Update LOINC loader to support legacy or modern CSVs
-The latest versions of LOINC were not working prior to this fix/update
as they contained a different format with data at different indexes in
multiple locations as well as extra data and missing data when compared.
The relevant fields have been accounted for and instead of a straight
fix to the latest version both modern and legacy versions are supported
in a dynamic fashion on load.