Track total validations run based on count in configured XML
-API provides a getter to the new "GlobalCodeValidatorResults" which
currently only contains a "vocabularyValidationConfigurationsCount."
vocabularyValidationConfigurationsCount is semi-dynamic in that it
tracks the total vocabulary validations configured for that particular
use (validation done) by the code validator. Since the configuration can
be dynamicaly swapped for each instance, a different count can be
tracked in this case. Otherwise, if the configuration is static, the
count would remain static for the duration of the referenceccdaservice
deployment.
--example of JSON in referenceccdaservice
---"vocabularyValidationConfigurationsCount":179
-Added basic testing